+++ example/openpose/pose-paf-master/vgg_model.py
@@ -1,15 +1,11 @@
 import tensorflow as tf
-from tensorlayer.layers import (Conv2d, ConcatLayer, DropoutLayer)
 import tensorlayer as tl
 from tensorlayer.layers import *
 
 __all__ = ['model']
-# _init_norm = tf.truncated_normal_initializer(stddev=0.01)
 _init_xavier = tf.contrib.layers.xavier_initializer()
 _init_norm = tf.truncated_normal_initializer(stddev=0.01)
 _b_initer = tf.constant_initializer(value=0.0)
-# _init_norm = tf.contrib.layers.xavier_initializer()
-# _b_initer = tf.contrib.layers.xavier_initializer()
 
 def _stage(cnn, b1, b2, n_pos,maskInput1,maskInput2, name='stageX'):
     with tf.variable_scope(name):
@@ -34,14 +30,9 @@ def _stage(cnn, b1, b2, n_pos,maskInput1,maskInput2, name='stageX'):
             b2.outputs = b2.outputs * maskInput2
     return b1, b2
 
-
 def vgg_network(x):
     red, green, blue = tf.split(axis=3, num_or_size_splits=3, value=x)
-    bgr = tf.concat(axis=3, values=[
-        blue ,
-        green ,
-        red 
-    ])
+    bgr = tf.concat(axis=3, values=[blue, green, red])
     bgr=bgr-0.5
     # input layer
     net_in = InputLayer(bgr, name='input')
@@ -75,7 +66,6 @@ def model(x, n_pos, mask_miss1,mask_miss2,is_train=False, reuse=None,):
     b2_list = []
     with tf.variable_scope('model', reuse):
         # feature extraction part
-        # cnn = tl.models.MobileNetV1(x, end_with='depth5', is_train=is_train, reuse=reuse)  # i.e. vgg16 conv4_2 ~ 4_4
         cnn = vgg_network(x)
         with tf.variable_scope('cpm', reuse):
             # stage 1
